requirements.txt
python-dotenv
fastapi
uvicorn 
pydantic
psycopg2
sqlalchemy
sqlmodel
strawberry-graphql[fastapi]
asyncpg
